Move `testController` to authentication_test.go
